feat: Add icon.types and icon.const
Introduce `icon.types` for managing component icon-related type definitions,
and `icon.const` for storing related constant assertions.
Subsequent updates have been made to all subscribers for seamless integration
with the new types and constant assertions.
