+++ components/icon/icon.const.ts
@@ -1,0 +1,5 @@
+export type VIEWBOX = (typeof VIEWBOX)[keyof typeof VIEWBOX];
+export const VIEWBOX = {
+  24: '0 0 24 24',
+  96: '0 96 960 960',
+} as const;
